add `Heap::allocate_raw` and `deallocate_raw`
This branch adds new APIs for lower-level manual memory management.
The `Heap::allocate_raw` method takes a user-provided `Layout` and
allocates an unsized, untyped allocation for that layout. The `unsafe`
free function `deallocate_raw` deallocates a pointer and `Layout` pair
that were allocated by `allocate_raw`. Together, these (unsafe) APIs
allow the implementation of user-defined types which manually manage a
dynamically-sized allocation, which was not previously possible with
`mnemos-alloc`.

Closes #4

Note that unlike the proposed interface in #4, this does *not* allow a
`HeapBox` to own a `!Sized` type. I'd like to figure that out
eventually, but it's somewhat orthogonal to the central concern of "how
do I allocate a raw `Layout`?