add missing `Send` impl for `AHeap` (#9)
Currently, any Future that awaits one of the AHeap allocate,
allocate_arc, allocate_array_with, or allocate_fixed_vec methods
is !Send. This is because these methods return a Future which
contains a &self receiver, and the AHeap type itself is !Send.
This is sad and not very great.

This branch adds a missing Send impl to AHeap. This should be safe
for the same reason that the existing unsafe impl Sync for AHeap is
safe.

* add test asserting heap futures are Send
* add missing Send impl for AHeap